Refactor main part of infrastructure-map into component
