+++ web/client/src/components/infrastructure/addIcons.js
