+++ web/client/src/components/infrastructure/mapStyle.js
